finished in-app cache and refactored code
